% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_inv_ALS_calc.R
\name{compete_ALS}
\alias{compete_ALS}
\title{Quantify Competition using inventory data from ALS}
\usage{
compete_ALS(seg_path, tree_path, radius)
}
\arguments{
\item{seg_path}{character path to inventory table (.csv or .txt) with structure: ID, X, Y, H}

\item{tree_path}{character path to table/list (.csv or .txt) of target trees within plot with ID_target, X, Y (does not have to be the same ID as in inventory table)}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all neighboring trees are classified as competitors}
}
\value{
dataframe with ID_target, CI12, CI13
}
\description{
'Compete_ALS()' returns a specific distance-dependent competition index (or group of indexes) for a target tree within a forest plot
}
\details{
Using an inventory table to easily quantify distance-dependent tree competition for row of trees within a plot.
The input data can either be taken directly from field measurements or derived beforehand from LiDAR point clouds.
It calculates two Competition indices, based on tree heights and distance to competitors.
}
\section{Methods}{

\itemize{
\item CI12 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} arctan(h_{i} / dist_{i})}
\item CI13 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (h_{i} / h) * arctan(h_{i} / dist_{i})}
}

Both indices are also described and compared with others in Contreras et al. (2011).

Tree Segmentation:

Various approaches can be used to segment airborne laser scanning point clouds into single trees and to obtain inventory data based it. Existing R packages for this are for example:
\itemize{
\item lidR package with different options to segment the point cloud or a Canopy Height Model (CHM)
\item itcLiDARallo within the package itcSegment
}

Be careful with low resolution/low density point clouds, as oversegmentation of trees is usually an issue!

Literature:
\itemize{
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree crowns in relation to local competition in a natural mature Scot pine forest. Can. J. For. Res. 27, 890–902.
\item Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree competition indices as predictors of basal area increment in western Montana forests. Forest Ecology and Management, 262(11): 1939-1949.
}
}

\examples{
\dontrun{
CI <- compete_ALS("path/to/invtable.csv", "path/to/target_trees.csv", radius = 10)
}
}
