% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_inv_ALS_calc.R
\name{compete_ALS_calc}
\alias{compete_ALS_calc}
\title{QUantify Competition using inventory data from ALS}
\usage{
compete_ALS_calc(path, radius = 10, dbh_thr = 0.1, target_tree, tolerance = 2)
}
\arguments{
\item{path}{character string path to folder with 1 or more.csv file(s) with inventory data with structure (ID, X, Y, DBH, H), DBH and H in m. Each row indicates one tree within the plot}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all neighboring trees are classified as competitors}

\item{dbh_thr}{numeric, DBH threshold for classifying the tree as a competitor (default is 0.1 m)}

\item{target_tree}{numeric (ID) or a vector of coordinates c(X, Y)}

\item{tolerance}{numeric. Tolerance for the match with the tree coordinates. If a field measurement value is used for target_tree, take a higher tolerance value (default=0.1 m), depending on the measurement accuracy}
}
\value{
numeric. Competition Index value
}
\description{
'compete_calc()' returns a specific distance-dependent competition index (or group of indexes) for a target tree within a forest plot
}
\details{
Using an inventory table to easily quantify distance-dependant tree competition for a single tree within a plot.
The input data can either be taken directly from field measurements or derived beforehand from LiDAR point clouds.
It is possible to choose between certain Competition indices, based on tree heights and distance to competitors
}
\section{Methods}{

\itemize{
\item CI11 according to Rouvinen & Kuuluvainen (1997)
\item CI12 according to Rouvinen & Kuuluvainen (1997)
\item CI13 according to Rouvinen & Kuuluvainen (1997)
...
}
}

\section{Literature}{

\itemize{
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree crowns in relation to local competition in a natural mature Scot pine forest. Can. J. For. Res. 27, 890–902.
}
}

\examples{
#......
}
