% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_cone.R
\name{competition_cone}
\alias{competition_cone}
\title{Quantify Tree Crown Competition (Cone Method)}
\usage{
competition_cone(
  forest_path,
  tree_path,
  comp_method = c("cone", "cylinder"),
  cylinder_r = 4
)
}
\arguments{
\item{forest_path}{character path to file of neighborhood point cloud (including the target tree and neighbors, not height normalized, can include ground points)}

\item{tree_path}{character path to file of target tree point cloud}

\item{comp_method}{character string with competition method "cone" or "cylinder"}

\item{cylinder_r}{(Optional) only important when using comp_method "cylinder"; numeric value of cylinder radius in m. Default is 4 m.}
}
\value{
vector of counted points and voxels of neighborhood point cloud that reach into the cone spanned over the target tree
}
\description{
Quantify Tree Crown Competition (Cone Method)
}
