% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{competition_indices}
\alias{competition_indices}
\alias{CI_Hegyi}
\alias{CI_Braathe}
\alias{CI_size}
\alias{CI_RK1}
\alias{CI_RK2}
\alias{CI_RK3}
\alias{CI_RK4}
\title{List of distance-based competition indices}
\usage{
CI_Hegyi(target, neigh)

CI_Braathe(target, neigh)

CI_size(target, neigh)

CI_RK1(target, neigh)

CI_RK2(target, neigh)

CI_RK3(target, neigh)

CI_RK4(target, neigh)
}
\arguments{
\item{target}{single row \code{forest_inv} class data.table object containing all
the variables needed to calculate the corresponding competition index for
the target tree (i.e. ´x´, ´y´, as well as ´dbh´, ´height´, ´size´ and/or
other size-related attributes).}

\item{neigh}{\code{forest_inv} class data.table object containing all the
variables needed to calculate the corresponding competition index for all
neighbor trees in the search radius of the target tree (i.e. ´x´, ´y´, as
well as´dbh´, ´height´, ´size´ and/or other size-related attributes), that
moreover includes their distance to the target tree in m (\code{dist}).}
}
\value{
A single numeric value (the competition index for the target tree).
}
\description{
list of distance-based competition index functions for the use
with \code{\link[=compete_inv]{compete_inv()}}.
}
\details{
\code{TreeCompR} provides a number of distance-based competition indices
for the use with \code{\link[=compete_inv]{compete_inv()}}.
\subsection{Available competition indices}{

The competition indices are computed according to the following
equations, where \eqn{d_i} and \eqn{h_i} are the dbh and height of neighbor
tree \eqn{i}, \eqn{d} and \eqn{h} are dbh and height of the target tree,
and \eqn{dist_i} is the distance from neighbor tree \eqn{i} to the target
tree.

\emph{Diameter-based competition indices}
\itemize{
\item CI_Hegyi introduced by Hegyi (1974): \cr
\eqn{CI_{Hegyi} = \sum_{i=1}^{n} d_{i} / (d \cdot dist_{i})}
\item CI_RK1 according to CI1 Rouvinen & Kuuluvainen (1997):\cr
\eqn{CI_{RK1} = \sum_{i=1}^{n} \mathrm{arctan}(d_{i} / dist_{i})}
\item CI_RK2 according to CI3 in Rouvinen & Kuuluvainen (1997): \cr
\eqn{CI_{RK2} =\sum_{i=1}^{n} (d_{i} / d) \cdot \mathrm{arctan}(d_{i}
   / dist_{i})}
}

\emph{Height-based competition indices}
\itemize{
\item CI_Braathe according to Braathe (1980): \cr
\eqn{CI_{Braathe} = \sum_{i=1}^{n} h_{i} / (h \cdot dist_{i})}
\item CI_RK3 according to CI5 in Rouvinen & Kuuluvainen (1997): \cr
\eqn{CI_{RK3} = \sum_{i=1}^{n} \mathrm{arctan}(h_{i} / dist_{i})}
for all trees with \eqn{h_{i} > h}
\item CI_RK4 based on CI3 in Rouvinen & Kuuluvainen (1997) and
Contreras et al. (2011): \cr
\eqn{CI_{RK4} = \sum_{i=1}^{n} (h_{i} / h) \cdot
   \mathrm{arctan}(h_{i} / dist_{i})}
}

\emph{Generic size-based Hegyi-type competition indices}
\itemize{
\item CI_size based on Hegyi (1974), but with a user-specified size-related
variable: \cr
\eqn{CI_{size} = \sum_{i=1}^{n} size_{i} / (size \cdot dist_{i})}
}
}

\subsection{User-specified competition indices}{

The creation of new user-specified competition indices is easily possible,
the only requirement is that they take two argument, \code{target} (a single-row
\code{forest_inv} class data.table with data for the target tree) and \code{neigh}
(a multi-row \code{forest_inv} class data.table with data for the neighbor
trees), and return a single numeric value. This function will then be called
internally for each tree and its neighborhood inside \code{compete_inv()}.
For example, a working implementation of the classical Hegyi index would
look like this:

\code{CI_Hegyi <- function(target, neigh) sum(neigh$dbh / (target$dbh * neigh$dist))}

It is advisable to add checks to ensure the function does the right thing
when it gets passed on empty datasets (e.g. because no trees are in the
neighborhood of a tree) because \code{sum(NULL)} evaluates to \code{0} which may not
always be intended.
}

\subsection{Literature}{
\itemize{
\item Hegyi, F., 1974. A simulation model for managing jackpine stands. In:
Fries, J. (Ed.), Proceedings of IUFRO meeting S4.01.04 on Growth models
for tree and stand simulation, Royal College of Forestry, Stockholm.
\item Braathe, P., 1980. Height increment of young single trees in relation to
height and distance of neighboring trees. Mitt. Forstl. VersAnst. 130,
43–48.
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree
crowns in relation to local competition in a natural mature Scot pine
forest. Can. J. For. Res. 27, 890–902.
\item Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree
competition indices as predictors of basal area increment in western
Montana forests. Forest Ecology and Management, 262(11): 1939-1949.
}
}
}
\examples{
\dontrun{
  # read data from all trees in a plot
  data <- read_inv("path/to/data.csv")

  # compute the Hegyi for the first tree in the dataset, assuming that all
  # others were in its search radius
  CI_Hegyi(data[1,], data[-1, ])
}
}
\seealso{
\code{\link[=read_inv]{read_inv()}} to read forest inventory data,
\code{\link[=define_target]{define_target()}} for designating target trees,
\code{\link[=compete_inv]{compete_inv()}} for computing tree competition from inventory data,
\code{\link[=plot_target]{plot_target()}} to plot target tree positions in \code{target_inv} and
\code{compete_inv} objects.
}
