% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_height.R
\name{height}
\alias{height}
\title{Tree Height of the Point Cloud}
\usage{
height(tree)
}
\arguments{
\item{tree}{data frame (X, Y, Z) of the tree point cloud}
}
\value{
tree height (numeric)
}
\description{
Tree Height of the Point Cloud
}
