% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_target.R
\name{plot_target}
\alias{plot_target}
\title{Plot selection of target trees}
\usage{
plot_target(inv, radius = NULL)
}
\arguments{
\item{inv}{object of class \code{compete_inv} or \code{target_inv}.}

\item{radius}{numeric of length 1, Search radius (in m) around target tree
wherein all neighboring trees are classified as competitors. Only needed
for \code{target_inv} objects in methods that are not radius dependent (i.e.,
in cases where \code{target_source} defined by a character or logical vector or
in cases where \code{target_source = "all_trees"}).}
}
\value{
A plot of the spatial arrangement of target trees.
}
\description{
\code{plot_target()} can be used with  a \code{target_inv} dataset or the output
of the \code{compete_inv} function to inspect the spatial positions of the
target trees.
}
\details{
The function creates a plot of the trees in the forest inventory
dataset where the target trees and the surrounding search radii are
highlighted. If they were created with \code{target_source = "buff_edge"} or
\code{target_source = "exclude_edge"}, the estimated plot margin (and, in case
of \code{"buff_edge"}, also the buffer to the margin) are added as a polygon.

This function is meant as a visual inspection tool for checking the
validity of the choice of target trees.
}
\examples{
\dontrun{
  # plot neighborhood for existing inventory
  comp <- compete_inv(inv, "buff_edge", radius = 10)
  plot_target(comp)
}

}
\seealso{
\code{\link[=read_inv]{read_inv()}} to read forest inventory data,
\code{\link[=define_target]{define_target()}} for designating target trees,
\code{\link[=compete_inv]{compete_inv()}} for computing tree competition from inventory data,
\link{competition_indices} for a list of available indices.
}
