% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_height.R
\name{position}
\alias{position}
\title{Position of the Tree from Point Cloud}
\usage{
position(tree)
}
\arguments{
\item{tree}{data frame (X, Y, Z) of the tree point cloud. Coordinates have to be in metric system!}
}
\value{
X,Y,Z coordinates of the tree's position
}
\description{
Position of the Tree from Point Cloud
}
\examples{
\dontrun{
# Read the tree point cloud
tree <- read_tree(path = "path/to/tree_point_cloud.txt")
# Get the position of this tree (X, Y, Z)
pos <- position(tree)
}
}
