% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Read a single tree point cloud}
\usage{
read_tree(path, ...)
}
\arguments{
\item{path}{character path to point cloud of individual tree or a whole plot
in las/laz format or file format readable with \code{\link[data.table:fread]{data.table::fread()}}}

\item{...}{additional arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
data frame with X,Y and Z coordinates of the tree or forest point
cloud
}
\description{
Read a single tree point cloud
}
\details{
Internal function for reading point cloud data. Currently, the
supported formats are las/laz and formats readable with
\code{\link[data.table:fread]{data.table::fread()}}. For other formats, please load the point cloud data
separately and enter the coordinates in downstream functions as a
data.frame.

If provided with tabular data, the function by default takes the columns
named "X", "Y", and "Z" or "x", "y", and "z" to be the coordinate vectors.
If no columns with matching names are available, it takes the first three
numeric columns and returns a message. If the dataset does not contain
three numeric columns or one of the columns labeled X, Y, and Z
is not numeric, the function fails with an error.
}
\examples{
\dontrun{
# Read the tree point cloud in txt format
tree <- read_tree(path = "path/to/tree_point_cloud.txt")
# Read the tree point cloud in las or laz format
tree <- read_tree(path = "path/to/tree_point_cloud.las")
}
}
