% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCompR-package.R
\docType{package}
\name{TreeCompR-package}
\alias{TreeCompR}
\alias{TreeCompR-package}
\title{TreeCompR: QUantifying individual tree competition}
\description{
Quantify individual tree competition using inventory data or point clouds.
}
\author{
\strong{Maintainer}: Julia Rieder \email{julia.rieder@uni-wuerzburg.de} (\href{https://orcid.org/0000-0001-7233-0488}{ORCID})

}
\keyword{internal}
