% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_inventory.R
\name{compete_calc}
\alias{compete_calc}
\title{Calculating distance-dependant Tree Competition Using Inventory Data (field or from LiDAR data)}
\usage{
compete_calc(
  path,
  radius = 10,
  dbh_thr = 0.1,
  target_tree,
  type = c("ID", "coordinates"),
  tolerance = 0.1,
  method = c("all", "Hegyi", "CI12", "CI13")
)
}
\arguments{
\item{path}{character string path to .csv file with inventory data with structure (ID, X, Y, DBH, H)}

\item{radius}{numeric, Search radius around target tree, wherein all neighboring trees are classified as competitors}

\item{dbh_thr}{numeric, DBH threshold for classifying the tree as a competitor (default is 0.1 m)}

\item{target_tree}{numeric (ID) or a vector of coordinates (X, Y)}

\item{type}{character string assigning the type of input of target_tree "ID" or "coordinates"}

\item{tolerance}{numeric. tolerance for matching the tree coordinates. If a field measured value is used for target_tree, take a higher tolerance value (default=0.1 m), depending on measurement accuracy}

\item{method}{character string assigning the method for quantifying competition "Hegyi", "CI12", "CI13"}
}
\value{
numeric. Competition Index value
}
\description{
Calculating distance-dependant Tree Competition Using Inventory Data (field or from LiDAR data)
}
\examples{
\dontrun{
# Calculate some distance-dependent CIs for one tree inside forest plot
# input coordinates target tree
ttree <- c(15, 9)
CI <- compete_calc("path/to/invtable.csv", dbh_thr = 0.1, ttree, "coordinates", 0.1, method = "all")

# Calculate the Hegyi-Index for one tree inside a forest plot, giving the ID of the target tree
ID_tree <- 5
CI <- compete_calc("path/to/invtable.csv", dbh_thr = 0.1, ID_tree, "ID", 0.1, method = "Hegyi")
#}
}
