% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_inventory.R
\name{compete_dd}
\alias{compete_dd}
\title{Calculate Tree Competition Indices From Inventory Data}
\usage{
compete_dd(
  path,
  radius = 10,
  dbh_thr = 0.1,
  target_tree,
  type = c("ID", "coordinates"),
  tolerance = 1,
  method = c("all", "Hegyi", "CI11", "CI12", "CI13")
)
}
\arguments{
\item{path}{character string path to .csv file with inventory data with structure (ID, X, Y, DBH, H), DBH and H in m. Coordinates have to be in metric system! Each row indicates one tree within the plot.}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all neighboring trees are classified as competitors}

\item{dbh_thr}{numeric, DBH threshold for classifying the tree as a competitor (default is 0.1 m)}

\item{target_tree}{numeric (ID) or a vector of coordinates c(X, Y)}

\item{type}{character string specifying the type of input of target_tree "ID" or "coordinates".}

\item{tolerance}{numeric. Tolerance for the match with the tree coordinates. If a field measurement value is used for target_tree, take a higher tolerance value (default=1 m), depending on the measurement accuracy}

\item{method}{character string assigning the method for quantifying competition "Hegyi", "CI10", CI11", "CI12", "CI13" or "all"}
}
\value{
numeric. Competition Index value
}
\description{
'compete_dd()' returns a specific distance-dependent competition index (or group of indexes) for a target tree within a forest plot
}
\details{
Using an inventory table to easily quantify distance-dependant tree competition for a single tree within a plot.
The input data can either be taken directly from field measurements or derived beforehand from LiDAR point clouds.
It is possible to choose between certain Competition indices, like the Hegyi index (method = "Hegyi") according to Hegyi (1974).
}
\section{Methods}{

\itemize{
\item Hegyi Index introduced by Hegyi (1974)
\eqn{\sum_{i=1}^{n} d_{i} / (d * dist_{i})}
\item CI10 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} arctan(d_{i} / dist_{i})}
\item CI11 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (d_{i} / d) * arctan(d_{i} / dist_{i})}
\item CI12 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} arctan(h_{i} / dist_{i})}
\item CI13 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (h_{i} / h) * arctan(h_{i} / dist_{i})}
}
}

\section{Literature}{

\itemize{
\item Hegyi, F., 1974. A simulation model for managing jackpine stands. In: Fries, J. (Ed.), Proceedings of IUFRO meeting S4.01.04 on Growth models for tree and stand simulation, Royal College of Forestry, Stockholm.
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree crowns in relation to local competition in a natural mature Scot pine forest. Can. J. For. Res. 27, 890–902.
\item Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree competition indices as predictors of basal area increment in western Montana forests. Forest Ecology and Management, 262(11): 1939-1949.
}
}

\examples{
\dontrun{
# Calculate some distance-dependent CIs for one tree inside forest plot
# input coordinates target tree
ttree <- c(15, 9)
CI <- compete_dd("path/to/invtable.csv", dbh_thr = 0.1, ttree, "coordinates", 1, method = "all")

# Calculate the Hegyi-Index for one tree inside a forest plot, giving the ID of the target tree
ID_tree <- 5
CI <- compete_dd("path/to/invtable.csv", dbh_thr = 0.1, ID_tree, "ID", 1, method = "Hegyi")
}
}
\seealso{
\code{\link[=competition_pc]{competition_pc()}} to quantify competition directly from point clouds, or \code{\link[=compete_dh]{compete_dh()}} if you do not have DBH data
}
