% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compete_inv.R
\name{compete_inv}
\alias{compete_inv}
\title{Calculate Tree Competition Indices From Inventory Data}
\usage{
compete_inv(
  seg_path,
  tree_path,
  radius,
  method = c("all", "Hegyi", "CI11", "CI12", "CI13")
)
}
\arguments{
\item{seg_path}{character path to inventory table (.csv or .txt) with structure: ID, X, Y, H}

\item{tree_path}{character path to table/list (.csv or .txt) of target trees within plot with ID_target, X, Y (does not have to be the same ID as in inventory table)}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all neighboring trees are classified as competitors}

\item{method}{character string assigning the method for quantifying competition "Hegyi", "CI11", "CI12", "CI13" or "all"}
}
\value{
dataframe with ID_target, and one or more indices per tree
}
\description{
'compete_inv()' returns a specific distance-dependent competition index (or group of indexes) for a list of target trees within a forest plot
}
\examples{
\dontrun{
CI <- compete_inv("path/to/invtable.csv", "path/to/target_trees.csv", radius = 10, method = "all")
}
}
