% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compete_inv.R
\name{compete_inv}
\alias{compete_inv}
\title{Calculate Tree Competition Indices From Inventory Data}
\usage{
compete_inv(
  seg_path,
  tree_path,
  radius,
  method = c("all", "Hegyi", "CI10", "CI11", "CI12", "CI13")
)
}
\arguments{
\item{seg_path}{character path to inventory table (.csv or .txt) with structure: ID, X, Y, H}

\item{tree_path}{character path to table/list (.csv or .txt) of target trees within plot with ID_target, X, Y (does not have to be the same ID as in inventory table)}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all neighboring trees are classified as competitors}

\item{method}{character string assigning the method for quantifying competition "Hegyi", "CI11", "CI12", "CI13" or "all"}
}
\value{
dataframe with ID_target, and one or more indices per tree, which depends on the chosen method
}
\description{
'compete_inv()' returns a specific distance-dependent competition index (or group of indexes) for a list of target trees within a forest plot
}
\section{Methods}{

\itemize{
\item Hegyi Index introduced by Hegyi (1974)
\eqn{\sum_{i=1}^{n} d_{i} / (d * dist_{i})}
\item CI10 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} arctan(d_{i} / dist_{i})}
\item CI11 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (d_{i} / d) * arctan(d_{i} / dist_{i})}
\item CI12 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} arctan(h_{i} / dist_{i})}
\item CI13 according to Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (h_{i} / h) * arctan(h_{i} / dist_{i})}
}
}

\section{Literature}{

\itemize{
\item Hegyi, F., 1974. A simulation model for managing jackpine stands. In: Fries, J. (Ed.), Proceedings of IUFRO meeting S4.01.04 on Growth models for tree and stand simulation, Royal College of Forestry, Stockholm.
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree crowns in relation to local competition in a natural mature Scot pine forest. Can. J. For. Res. 27, 890–902.
\item Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree competition indices as predictors of basal area increment in western Montana forests. Forest Ecology and Management, 262(11): 1939-1949.
}
}

\examples{
\dontrun{
CI <- compete_inv("path/to/invtable.csv", "path/to/target_trees.csv", radius = 10, method = "all")
}
}
