% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compete_inventory.R
\name{compete_inv}
\alias{compete_inv}
\title{Quantify distance-dependent Competition using inventory data}
\usage{
compete_inv(
  inv_source,
  target_source,
  radius,
  method = c("all", "CI_Hegyi", "CI_Braathe", "CI_RK1", "CI_RK2", "CI_RK3", "CI_RK4"),
  x = NULL,
  y = NULL,
  dbh = NULL,
  height = NULL,
  id = NULL,
  dbh_unit = c("cm", "m", "mm"),
  height_unit = c("m", "cm", "mm"),
  verbose = TRUE,
  tolerance = 1,
  ...
)
}
\arguments{
\item{inv_source}{dataframe or path to inventory table of the plot, with
structure: ID, x, y, dbh (in cm) and/or h (in m). Cartesian coordinates have
to be in metric system (in m)!}

\item{target_source}{dataframe or path to table of target trees within plot
with ID_target, x, y (does not have to be the same ID as in inventory table).
Cartesian coordinates have to be in metric system!
(and same coordinate system as inv_source!)}

\item{radius}{numeric, Search radius (in m) around target tree, wherein all
neighboring trees are classified as competitors}

\item{method}{character string assigning the method for quantifying
competition dbh-distance-dependent: "CI_Hegyi", "CI_RK1", "CI_RK2",
height-distance-dependent: "CI_Braathe", "CI_RK3", "CI_RK4" or "all"
for all the indices}

\item{tolerance}{numeric. Tolerance for the match with the tree coordinates.
If coordinates are measured in the field with GPS, but inv_source contains
x,y from segmentation, allow for tolerance to match the target tree
positions. (default = 1 m), depending on the measurement accuracy of GPS}

\item{...}{additional arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
dataframe with ID_target, and one or more Indices depending on
chosen method
}
\description{
'compete_inv()' returns a specific distance-height-dependent or
distance-dbh-dependent competition index (or group of indices) for all trees
within a forest plot or specified target trees
}
\details{
Using an inventory table to easily quantify distance-dependent tree
competition for a list of trees within a plot or all the trees.
The input data can either be taken directly from field measurements or
derived beforehand from LiDAR point clouds.
The function calculates 6 different Competition indices, based on tree
heights or dbh and distance to competitors.
}
\section{Methods}{

\itemize{
\item CI_Hegyi Index introduced by Hegyi (1974)
\eqn{\sum_{i=1}^{n} d_{i} / (d \cdot dist_{i})}
\item CI_RK1 according to CI1 Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} \mathrm{arctan}(d_{i} / dist_{i})}
\item CI_RK2 according to CI3 in Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} (d_{i} / d) \cdot \mathrm{arctan}(d_{i} / dist_{i})}
\item CI_Braathe according to Braathe (1980)
\eqn{\sum_{i=1}^{n} h_{i} / (h \cdot dist_{i})}
\item CI_RK3 according to CI5 in Rouvinen & Kuuluvainen (1997)
\eqn{\sum_{i=1}^{n} \mathrm{arctan}(h_{i} / dist_{i}), h_{i} > h}
\item CI_RK4 based on CI3 in Rouvinen & Kuuluvainen (1997) and
Contreras et al. (2011)
\eqn{\sum_{i=1}^{n} (h_{i} / h) \cdot \mathrm{arctan}(h_{i} / dist_{i})}
}
}

\section{Tree Segmentation}{

Various approaches can be used to segment (airborne) laser scanning point
clouds into single trees and to obtain inventory data based it. Existing R
packages for this are for example:
\itemize{
\item TreeLS package for automated segmentation of terrestrial/mobile laser scans
\item lidR package with different options to segment the point cloud or a Canopy
Height Model (CHM)
\item itcLiDARallo within the package itcSegment
}

Be careful with low resolution/low density point clouds, as oversegmentation
of trees is usually an issue!
}

\section{Literature}{

\itemize{
\item Hegyi, F., 1974. A simulation model for managing jackpine stands. In:
Fries, J. (Ed.), Proceedings of IUFRO meeting S4.01.04 on Growth models
for tree and stand simulation, Royal College of Forestry, Stockholm.
\item Braathe, P., 1980. Height increment of young single trees in relation to
height and distance of neighboring trees. Mitt. Forstl. VersAnst. 130,
43–48.
\item Rouvinen, S., Kuuluvainen, T., 1997. Structure and asymmetry of tree
crowns in relation to local competition in a natural mature Scot pine
forest. Can. J. For. Res. 27, 890–902.
\item Contreras, M.A., Affleck, D. & Chung, W., 2011. Evaluating tree
competition indices as predictors of basal area increment in western
Montana forests. Forest Ecology and Management, 262(11): 1939-1949.
}
}

\examples{
\dontrun{
CI <- compete_inv("path/to/invtable.csv",
  "path/to/target_trees.csv", radius = 10, method = "CI_Hegyi")
}
}
