% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_cone.R
\name{competition_pc}
\alias{competition_pc}
\title{Quantify Tree Crown Competition (Cone Method or cylinder method)}
\usage{
competition_pc(
  forest_path,
  tree_path,
  comp_method = c("cone", "cylinder"),
  cyl_r = 4
)
}
\arguments{
\item{forest_path}{character path to file of neighborhood point cloud (including the target tree and neighbors, not height normalized, can include ground points)}

\item{tree_path}{character path to file of target tree point cloud}

\item{comp_method}{character string with competition method "cone" or "cylinder"}

\item{cyl_r}{(Optional) only important when using comp_method "cylinder"; numeric value of cylinder radius in m. Default is 4 m.}
}
\value{
vector of counted points and voxels of neighborhood point cloud that reach into the cone spanned over the target tree
}
\description{
opening angle 60°, spanned in 50 \% of target tree's height
}
\examples{
\dontrun{
#Quantify crown competition for a single tree within a plot using cone method
CI_cone <- competition_pc("path/to/forest_pc.txt", "path/to/tree_pc.txt", "cone")

#Quantify competition for a single tree within a plot using cylinder method with radius 4 m
CI_cyl <- competition_pc("path/to/forest_pc.txt", "path/to/tree_pc.txt", "cylinder", cyl_r = 4)
}
}
