% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_cone.R
\name{competition_pc}
\alias{competition_pc}
\title{Quantify Tree Competition from Point Clouds (Cone or cylinder method)}
\usage{
competition_pc(
  forest_path,
  tree_path,
  comp_method = c("cone", "cylinder"),
  cyl_r = 4
)
}
\arguments{
\item{forest_path}{character path to file of neighborhood point cloud (including the target tree and neighbors, not height normalized, can include ground points) Coordinates have to be in metric system!}

\item{tree_path}{character path to file of target tree point cloud. Coordinates have to be in metric system!}

\item{comp_method}{character string with competition method "cone" or "cylinder"}

\item{cyl_r}{(Optional) only important when using comp_method "cylinder"; numeric value of cylinder radius in m. Default is 4 m.}
}
\value{
vector of counted points and voxels of neighborhood point cloud that reach into the cone spanned over the target tree
}
\description{
Counting the points or voxels of neighboring trees that reach into search cone or cylinder for the target tree
}
\details{
\itemize{
\item Cone Method: search cone with opening angle 60°, spanned in 50 \% of target tree's height, counting the points and voxels of neighboring trees that reach within
\item Cylinder Method: search cylinder with specific radius (target tree in center), counting the points and voxels of neighboring trees that reach within
}

Check the Literature:
\itemize{
\item Metz, J., Seidel, D., Schall, P., Scheffer, D., Schulze, E.-D. & Ammer, C. (2013). Crown modeling by terrestrial laser scanning
as an approach to assess the effect of aboveground intra- and interspecific competition on tree growth. Forest Ecology and Management,
310: 275-288. https://doi.org/10.1016/j.foreco.2013.08.014
\item Seidel, D., Hoffmann, N., Ehbrecht, M., Juchheim, J. & Ammer, C. (2015). How neighborhood affects tree diameter increment – New insights from
terrestrial laser scanning and some methodical considerations. Forest Ecology and Management, 336: 119-128. http://dx.doi.org/10.1016/j.foreco.2014.10.020
}
}
\examples{
\dontrun{
#Quantify crown competition for a single tree within a plot using cone method
CI_cone <- competition_pc("path/to/forest_pc.txt", "path/to/tree_pc.txt", "cone")

#Quantify competition for a single tree within a plot using cylinder method with radius 4 m
CI_cyl <- competition_pc("path/to/forest_pc.txt", "path/to/tree_pc.txt", "cylinder", cyl_r = 4)
}
}
