% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inventory.R
\name{read_inv}
\alias{read_inv}
\alias{print.forest_inv}
\title{Read forest inventory data}
\usage{
read_inv(
  inv_source,
  x = NULL,
  y = NULL,
  dbh = NULL,
  height = NULL,
  id = NULL,
  dbh_unit = c("cm", "m", "mm"),
  height_unit = c("m", "cm", "mm"),
  verbose = TRUE,
  names_as_is = FALSE,
  ...
)
}
\arguments{
\item{inv_source}{object that inherits from class data.frame, or character
path to a file in any file format readable with \code{\link[data.table:fread]{data.table::fread()}}.
If provided with forest inventory data in a data.frame, the structure and
column names are validated and homogenized; else, the function tries to
read the dataset in the specified path.}

\item{x}{character of length 1 or name of the variable in \code{inv_source}
containing the x coordinates of the tree in m. If \code{NULL} (default),
the function tries to identify the x coordinate from the data.}

\item{y}{character of length 1 or name of the variable in \code{inv_source}
containing the y coordinates of the tree in meters. If \code{NULL} (default),
the function tries to identify the y coordinate from the data.}

\item{dbh}{character of length 1 or name of the variable in \code{inv_source}
containing the diameter at breast height of the tree (by default in cm, but
can be defined via \code{dbh_unit}). If \code{NULL} (default), the function tries to
identify the dbh from the data.}

\item{height}{character of length 1 or name of the variable in \code{inv_source}
containing the height of the tree (by default in m, but   can be defined
via \code{heigh_unit}). If \code{NULL} (default), the function tries to identify the
height from the data.}

\item{id}{character of length 1 or name of the variable in \code{inv_source}
containing a unique tree ID. If \code{NULL} (default), the function tries to
identify the ID from the data. If this is not possible, the trees are
assigned a unique number. All IDs are coerced to character.}

\item{dbh_unit}{character of length 1. Unit for the diameter measurements
(one of "cm", "m" or "mm". defaults to "cm").}

\item{height_unit}{character of length 1. Unit for the diameter measurements
(one of "m", "cm" or "mm". defaults to "m").}

\item{verbose}{logical of length 1. Should information about progress be
printed? Defaults to TRUE.}

\item{names_as_is}{logical of length 1. If TRUE, only NULL or characters
are excepted as column names in \code{x}, \code{y}, \code{dbh}, \code{height} and \code{id}, and
the column names are not internally substituted. Defaults to FALSE, which
should normally be the right choice for interactive use (TRUE is only
needed for use within functions).}

\item{...}{additional arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
object of class \code{forest_inv} with x and y coordinates of the tree, a
unique tree identifier (\code{id}) and tree diameter at breast height (\code{dbh}, in
cm) and tree height (\code{height}, in m) if available.
}
\description{
Read inventory table sourced from a file path or an object that
inherits from class data.frame.
}
\details{
Function for reading and validating forest inventory data.
Supports any formats readable with \code{\link[data.table:fread]{data.table::fread()}}.

If provided with tabular data without explicitly specified variable names,
the function by default takes the columns named "X" and "Y" (or "x" and
"y") to be the tree coordinates, and looks for columns named "height",
"height_m" or "h" as well as "dbh", "diameter","diam", or "d" (in any
capitalization) as size-related variables. The tree ids are taken from
columns named "id", "tree_id", "treeID" or "tree.id" (in any
capitalization). All special characters besides "." and "_" are stripped
from the column names before matching.

If no columns with coordinates can be identified, the function fails with
an error. If no ID column is available, the function assigns a unique
number to each tree (but note that this will make specifying custom target
trees difficult). It is possible to read in datasets without dbh and
height, but usually only sensible if these are used as accessory datasets
for identifying target trees (e.g. if target trees where identified by
taking their GPS coordinates manually in the field).
}
\examples{
\dontrun{
# prepare inventory table for compete_inv()
inv <- fread(path = "path/to/table.csv")
#specify the units of parameters within your input
inv_table <- read_inv(inv, dbh_unit = "cm", height_unit = "m")
# Read inventory table directly from directory
inv_table <- read_inv(inv_source = "path/to/table.csv", dbh_unit = "cm",
height_unit = "m")
}
}
\seealso{
\code{\link[=define_target]{define_target()}} for designating target trees,
\code{\link[=compete_inv]{compete_inv()}} for computing tree competition from inventory data,
\code{\link[=plot_target]{plot_target()}} to plot target tree positions in \code{target_inv} and
\code{compete_inv} objects.
}
