% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inventory.R
\name{read_inv}
\alias{read_inv}
\title{Read forest inventory data}
\usage{
read_inv(
  inv_source,
  x = NULL,
  y = NULL,
  dbh = NULL,
  height = NULL,
  id = NULL,
  dbh_unit = c("cm", "m", "mm"),
  height_unit = c("m", "cm", "mm"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{inv_source}{object that inherits from class data.frame, or character
path to a file in any file format readable with \code{\link[data.table:fread]{data.table::fread()}}.
If provided with forest inventory data in a data.frame, the structure and
column names are validated and homogenized; else, the function tries to
read the dataset in the specified path.}

\item{x}{character of length 1 or name of the variable in \code{inv_source}
containing the x coordinates of the tree in meters . If `NULL (default),
the function tries to identify the x coordinate from the data.}

\item{y}{character of length 1 or name of the variable in \code{inv_source}
containing the y coordinates of the tree in meters. If `NULL (default), the
function tries to identify the y coordinate from the data.}

\item{dbh}{character of length 1 or name of the variable in \code{inv_source}
containing the diameter at breast height of the tree (by default in cm, but
can be defined via \code{dbh_unit}). If \code{NULL} (default), the function tries to
identify the dbh from the data. At least one of \code{dbh} or \code{height} has to be
specified.}

\item{height}{character of length 1 or name of the variable in \code{inv_source}
containing the height of the tree (by default in m, but   can be defined
via \code{h_unit}). If \code{NULL} (default), the function tries to identify the dbh
from the data. At least one of \code{dbh} or \code{height} has to be specified.}

\item{id}{character of length 1 or name of the variable in \code{inv_source}
containing a unique tree ID. If \code{NULL} (default), the function tries to
identify the ID from the data. If this is not possible, a numeric idenfier
is assigned by the function.}

\item{dbh_unit}{character of length 1. Unit for the diameter measurements
(one of "cm", "m" or "mm". defaults to "cm").}

\item{height_unit}{character of length 1. Unit for the diameter measurements
(one of "m", "cm" or "mm". defaults to "m").}

\item{verbose}{logical of length 1. Should information about progress be
printed? Defaults to TRUE.}

\item{...}{additional arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
object of class \code{c("forest_inv", "data.frame")} with x and y
coordinates of the tree, a unique tree identifier (\code{ID}) and at least one
of tree diameter at breast height (\code{dbh}, in cm) and tree height (\code{height}, in
m).
}
\description{
Read point cloud sourced from a file path or an object that
inherits from class data.frame.
}
\details{
Function for reading and validating forest inventory data.
Supports any formats readable with \code{\link[data.table:fread]{data.table::fread()}}.

If provided with tabular data without explicitly specified variable names,
the function by default takes the columns named "X" and "Y" (or "x" and
"y") to be the tree coordinates, and looks for columns named
"height" or "h" as well as "dbh", "diameter","diam", or "d" (in any
capitalization) as size-related variables. The coordinates are taken from
columns named "id", "tree_id", "treeID" or "tree.id" (in any
capitalization).

If no columns with coordinates and/or none of either dbh or h can be
identified, the function fails with an error. If no ID column is available,
the function assigns a unique number to each tree (but note that this will
make specifying custom target trees difficult).
}
\examples{
\dontrun{
# Read the tree point cloud in txt format
tree <- read_tree(path = "path/to/tree_point_cloud.txt")
# Read the tree point cloud in las or laz format
tree <- read_tree(path = "path/to/tree_point_cloud.las")
}
}
