% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Read a single tree point cloud}
\usage{
read_tree(path)
}
\arguments{
\item{path}{character path to point cloud of individual tree or a whole plot
in txt or las/laz format.}
}
\value{
data frame with X,Y,Z of the tree or forest point cloud
}
\description{
Read a single tree point cloud
}
\details{
Internal function for reading point cloud data. Currently, the
supported formats are txt and las/laz. For other formats, please load the
point cloud data separately and enter the coordinates as a data.frame.
}
\examples{
\dontrun{
# Read the tree point cloud in txt format
tree <- read_tree(path = "path/to/tree_point_cloud.txt")
# Read the tree point cloud in las or laz format
tree <- read_tree(path = "path/to/tree_point_cloud.las")
}
}
