% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Read a single tree point cloud}
\usage{
read_tree(path)
}
\arguments{
\item{path}{character path to point cloud of individual tree or a whole plot}
}
\value{
data frame with X,Y,Z of the tree point cloud
}
\description{
Read a single tree point cloud
}
