% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeCompR-package.R
\docType{package}
\name{TreeCompR-package}
\alias{TreeCompR}
\alias{TreeCompR-package}
\title{TreeCompR: Tree Competition Indices For Inventory Data And 3D Point Clouds}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Quantify individual tree competition using inventory data or point clouds.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/juliarieder/TreeCompR/issues}
}

}
\author{
\strong{Maintainer}: Julia S. Rieder \email{julia.rieder@uni-wuerzburg.de} (\href{https://orcid.org/0000-0001-7233-0488}{ORCID})

Authors:
\itemize{
  \item Roman M. Link \email{roman.link@plant-ecology.de} (\href{https://orcid.org/0000-0003-0588-3757}{ORCID})
}

}
\keyword{internal}
